Move RepoDetails to Export/repositories/common
